Separate calc.py into calc.py and preprocess.py
